Create Arduino code for human following robot
This project is about the design and implementation of a human-following robot using Arduino Uno. The robot is equipped with IR sensors, ultrasonic sensors, and an L298N motor driver to accurately detect and follow humans. The software used to control the robot is written in Arduino programming language, which allows for easy customization and modification. The simulation of the robot's movement is done using Tinkercad a 3D design and Proteus 8 a simulation tool. The design of the robot is based on a simple chassis made from acrylic sheets, which is lightweight and easy to assemble. The robot is powered by a 9V battery, which provides enough power to run the motor and sensors for several hours. Overall, this project demonstrates the potential of robotics in everyday life and could inspire further research and development in this area.